test(integration): refactor
