+++ src/test/java/kr/hhplus/be/server/interfaces/api/order/OrderControllerTest.kt
@@ -1,10 +1,9 @@
 package kr.hhplus.be.server.interfaces.api.order
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import kr.hhplus.be.server.IntegrationTestSupport
-import kr.hhplus.be.server.LongFixture
 import kr.hhplus.be.server.domain.auth.UserId
 import kr.hhplus.be.server.domain.point.UserPoint
+import kr.hhplus.be.server.integration.IntegrationTestSupport
 import org.hamcrest.Matchers.`is`
 import org.instancio.Instancio
 import org.instancio.Select.field
@@ -25,7 +24,6 @@ class OrderControllerTest : IntegrationTestSupport() {
 
     @Autowired
     private lateinit var objectMapper: ObjectMapper
-    val longFixture = LongFixture()
     val userId = longFixture.userId()
     val orderId = longFixture.orderId()
