+++ src/test/java/kr/hhplus/be/server/interfaces/api/product/ProductControllerTest.kt
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.interfaces.api.product
 
-import kr.hhplus.be.server.IntegrationTestSupport
+import kr.hhplus.be.server.integration.IntegrationTestSupport
 import org.hamcrest.Matchers
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
